de-project-5/requirements.txt
kafka-python
pyyaml
pyspark